Restructure API-related code
