+++ include/zenoh-pico/api/logger.h
@@ -12,20 +12,14 @@
  *   ADLINK zenoh team, <zenoh@adlink-labs.tech>
  */
 
-#ifndef ZENOH_PICO_H
-#define ZENOH_PICO_H
+#ifndef ZENOH_PICO_LOGGER_API_H
+#define ZENOH_PICO_LOGGER_API_H
 
-#include "zenoh-pico/api/logger.h"
-#include "zenoh-pico/api/config.h"
-#include "zenoh-pico/api/session.h"
-#include "zenoh-pico/api/primitives.h"
-#include "zenoh-pico/api/resource.h"
-#include "zenoh-pico/api/query.h"
-#include "zenoh-pico/api/subscribe.h"
-#include "zenoh-pico/api/memory.h"
+#include "zenoh-pico/utils/logging.h"
 
-#if defined(ZENOH_ZEPHYR)
-#include <zephyr.h>
-#endif
+/**
+ * Initialise the zenoh runtime logger
+ */
+void z_init_logger(void);
 
-#endif /* ZENOH_PICO_H */
+#endif /* ZENOH_PICO_LOGGER_API_H */
