+++ src/api/subscribe.c
@@ -12,20 +12,13 @@
  *   ADLINK zenoh team, <zenoh@adlink-labs.tech>
  */
 
-#ifndef ZENOH_PICO_H
-#define ZENOH_PICO_H
-
-#include "zenoh-pico/api/logger.h"
-#include "zenoh-pico/api/config.h"
-#include "zenoh-pico/api/session.h"
-#include "zenoh-pico/api/primitives.h"
-#include "zenoh-pico/api/resource.h"
-#include "zenoh-pico/api/query.h"
 #include "zenoh-pico/api/subscribe.h"
-#include "zenoh-pico/api/memory.h"
 
-#if defined(ZENOH_ZEPHYR)
-#include <zephyr.h>
-#endif
-
-#endif /* ZENOH_PICO_H */
+zn_subinfo_t zn_subinfo_default()
+{
+    zn_subinfo_t si;
+    si.reliability = zn_reliability_t_RELIABLE;
+    si.mode = zn_submode_t_PUSH;
+    si.period = NULL;
+    return si;
+}
